Moved validation out of extractor, fixed tests
