+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/controllers/PublicationController.java
@@ -105,6 +105,7 @@ public class PublicationController {
                                                      displayFrom, displayTo, listType, courtId, contentDate);
 
         HeaderGroup headers = validationService.validateHeaders(initialHeaders);
+        validationService.validateBody(payload, initialHeaders.getListType());
 
         Artefact artefact = Artefact.builder()
             .provenance(headers.getProvenance())
