+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/utils/JsonExtractor.java
@@ -1,25 +1,15 @@
 package uk.gov.hmcts.reform.pip.data.management.utils;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.DocumentContext;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.Option;
-import com.networknt.schema.JsonSchema;
-import com.networknt.schema.JsonSchemaFactory;
-import com.networknt.schema.SpecVersion;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.pip.data.management.config.SearchConfiguration;
-import uk.gov.hmcts.reform.pip.data.management.config.ValidationConfiguration;
-import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.PayloadValidationException;
-import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
-import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -33,31 +23,14 @@ public class JsonExtractor implements Extractor {
 
     private final SearchConfiguration searchConfiguration;
 
-    private final JsonSchema masterSchema;
-    private final JsonSchema dailyCauseListSchema;
-
     @Autowired
-    public JsonExtractor(SearchConfiguration searchConfiguration, ValidationConfiguration validationConfiguration) {
+    public JsonExtractor(SearchConfiguration searchConfiguration) {
         this.searchConfiguration = searchConfiguration;
 
         jsonConfiguration = Configuration.defaultConfiguration()
             .addOptions(Option.DEFAULT_PATH_LEAF_TO_NULL)
             .addOptions(Option.SUPPRESS_EXCEPTIONS)
             .addOptions(Option.ALWAYS_RETURN_LIST);
-
-        try (InputStream masterFile = this.getClass().getClassLoader()
-            .getResourceAsStream(validationConfiguration.getMasterSchema());
-
-             InputStream dailyCauseListFile = this.getClass().getClassLoader()
-                 .getResourceAsStream(validationConfiguration.getDailyCauseList())) {
-
-            JsonSchemaFactory schemaFactory = JsonSchemaFactory.getInstance(SpecVersion.VersionFlag.V7);
-            masterSchema = schemaFactory.getSchema(masterFile);
-            dailyCauseListSchema = schemaFactory.getSchema(dailyCauseListFile);
-
-        } catch (Exception exception) {
-            throw new PayloadValidationException(String.join(exception.getMessage()));
-        }
 
     }
 
@@ -87,25 +60,6 @@ public class JsonExtractor implements Extractor {
             return true;
         } catch (IOException exception) {
             return false;
-        }
-    }
-
-    @Override
-    public List<String> validate(Artefact artefact, String payload) {
-        try {
-            List<String> errors = new ArrayList<>();
-
-            JsonNode json = new ObjectMapper().readTree(payload);
-            masterSchema.validate(json).forEach(vm ->  errors.add(vm.getMessage()));
-
-            if (artefact.getListType() != null && artefact.getListType().equals(ListType.CIVIL_DAILY_CAUSE_LIST)) {
-                dailyCauseListSchema.validate(json).forEach(vm ->  errors.add(vm.getMessage()));
-            }
-
-            return errors;
-
-        } catch (IOException exception) {
-            throw new PayloadValidationException("Error while reading JSON Payload");
         }
     }
