+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/utils/PayloadExtractorTest.java
@@ -5,15 +5,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.PayloadValidationException;
-import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -40,12 +37,10 @@ class PayloadExtractorTest {
     void testValidExtractorFound() {
         PayloadExtractor payloadExtractor = new PayloadExtractor(List.of(jsonExtractor));
 
-        Artefact artefact = new Artefact();
         when(jsonExtractor.isAccepted(PAYLOAD)).thenReturn(true);
-        when(jsonExtractor.validate(artefact, PAYLOAD)).thenReturn(List.of());
         when(jsonExtractor.extractSearchTerms(PAYLOAD)).thenReturn(TEST_MAP);
 
-        Map<String, List<Object>> searchTerms = payloadExtractor.validateAndParsePayload(artefact, PAYLOAD);
+        Map<String, List<Object>> searchTerms = payloadExtractor.extractSearchTerms(PAYLOAD);
         assertEquals(TEST_MAP, searchTerms, "Returned search terms does not match expected terms");
     }
 
@@ -53,14 +48,12 @@ class PayloadExtractorTest {
     void testMultipleExtractorsCalledTwice() {
         PayloadExtractor payloadExtractor = new PayloadExtractor(List.of(jsonExtractorOther, jsonExtractor));
 
-        Artefact artefact = new Artefact();
         when(jsonExtractor.isAccepted(PAYLOAD)).thenReturn(true);
-        when(jsonExtractor.validate(artefact, PAYLOAD)).thenReturn(List.of());
         when(jsonExtractor.extractSearchTerms(PAYLOAD)).thenReturn(TEST_MAP);
 
         when(jsonExtractorOther.isAccepted(PAYLOAD)).thenReturn(false);
 
-        Map<String, List<Object>> searchTerms = payloadExtractor.validateAndParsePayload(artefact, PAYLOAD);
+        Map<String, List<Object>> searchTerms = payloadExtractor.extractSearchTerms(PAYLOAD);
         verify(jsonExtractorOther, times(1)).isAccepted(PAYLOAD);
         assertEquals(TEST_MAP, searchTerms, "Returned search terms does not match expected terms");
     }
@@ -71,26 +64,8 @@ class PayloadExtractorTest {
         when(jsonExtractor.isAccepted(PAYLOAD)).thenReturn(false);
 
 
-        assertTrue(payloadExtractor.validateAndParsePayload(new Artefact(), PAYLOAD).isEmpty(),
+        assertTrue(payloadExtractor.extractSearchTerms(PAYLOAD).isEmpty(),
              "Returned search terms is not empty when no extract found");
-    }
-
-    @Test
-    void testExtractorFoundButInvalid() {
-        final String invalidPayload = "Payload invalid";
-        PayloadExtractor payloadExtractor = new PayloadExtractor(List.of(jsonExtractor));
-
-        Artefact artefact = new Artefact();
-        when(jsonExtractor.isAccepted(PAYLOAD)).thenReturn(true);
-        when(jsonExtractor.validate(artefact, PAYLOAD)).thenReturn(List.of(invalidPayload));
-
-        PayloadValidationException payloadValidationException =
-            assertThrows(PayloadValidationException.class, () ->
-                payloadExtractor.validateAndParsePayload(artefact, PAYLOAD));
-
-        assertTrue(
-            payloadValidationException.getMessage().contains(invalidPayload),
-            "Invalid error message is thrown");
     }
 
 }
